+++ src/components/Prescription/style.module.css
@@ -1,3 +1,10 @@
+@media screen and (min-width: 80rem) {
+  .container {
+    margin: 1em 2em;
+    max-width: 45%;
+    max-height: var(45);
+  }
+}
 .prescriptionBox {
   display: flex;
   flex-direction: column;
