+++ src/components/homePageComponents/aboutus/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import style from "./style.module.css";
-import { motion} from "framer-motion";
+import {motion} from "framer-motion";
 const Animate = {
     offscreen: {
         x: -10,
