+++ src/components/homePageComponents/aboutus/style.module.css
@@ -1,3 +1,4 @@
+
 .mainDiv{
     margin-top: 200px;
 }
@@ -25,6 +26,11 @@
     height: 100%;
     font-size: 1.2rem;
     color: var(--primary-color)
+}
+@media screen and (min-width:1171px){
+    .logo{
+        left:-90px;
+    }
 }
 .logo{
     height: 500px;
