+++ src/components/homePageComponents/indexpage/style.module.css
@@ -1,3 +1,10 @@
+@media screen and (min-width: 80rem) {
+    .container {
+      margin: 1em 2em;
+      max-width: 900px;
+      max-height: fit-content;
+    }
+  }
 .navbar{
     display: flex;
     justify-content: space-between;
